+++ package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "start": "node bin/www",
     "test": "jest",
-    "faker": " node fakers/script.faker.js ",
+    "faker": " node fakers/product.faker.js ",
     "swagger": "node swagger.js"
   },
   "author": "sushi-team",
